// mobile/GreenwichApp/src/resources/types.ts
export type TChannelValue = {
  badTerminal?: boolean;
  digitalStatus?: number;
  engineeringValue?: number;
};

export type TInputPointStatusResponse = {
  inputPointID: number;
  enabled: boolean;
  name: string;
  status: number;
};

export type TLiftChannel = {
  modbusChannelId: number;
  channelName: string;
  liftNumber: string;
  signalType: string;
  locationMask: number;
};

export type TLiftChannelData = {
  [liftName: string]: TLiftChannel;
};

export type TLiftSignalConversion = {
  value: number;
  displayValue: string;
}[];
// ...
export type TRootStackParamList = {
  Home: undefined;
  CCTV: undefined;
  DoorAccess: undefined;
  Electricity: undefined;
  Water: undefined;
  Leakage: undefined;
  AirFlow: undefined;
  WaterTank: undefined;
  Valve: undefined;
  Door: undefined;
  EmergencyAlarm: undefined;
  Lift: undefined;
  Settings: undefined;
};